:recycle: Move tests about filtering to a different file
